refactor: changed due to useValidateBudgetModal
